Order games by updated_at DESC by default; touch game when saving approved review
and don’t automatically approve a review when editing it as an admin.

This will cause updated and newly-reviewed games to jump to the top of
the list on the front page and games index.

This code should be run in a rails console to get all current games
into order based on when they were last reviewed:
PlayerReview.reorder('updated_at ASC').all.each { |r| r.save }